+++ bricks/clean_architecture_skeleton/__brick__/lib/shared/data/data_sources/profile_remote_data_source.dart
@@ -1,9 +1,12 @@
 import 'package:injectable/injectable.dart';
 
+import '../../../core/constants/json_constant.dart';
 import '../../../core/di/service_locator.dart';
 import '../../../core/network/api_endpoint.dart';
 import '../../../core/network/http/modules/{{name.snakeCase()}}_http_module.dart';
 
+import '../dtos/profile_dto.dart';
+import '../mapper/profile_mapper.dart';
 import '../models/profile_model.dart';
 
 abstract class ProfileRemoteDataSource {
@@ -17,6 +20,11 @@ class ProfileRemoteDataSourceImpl implements ProfileRemoteDataSource {
   @override
   Future<ProfileModel> getProfile() async {
     var response = await _httpClient.get(ApiEndpoint.baseUrl);
+
+    if(response.containsKey(JsonConstant.attributes)) {
+      var profileDto = ProfileDto.fromJson(response); 
+      return ProfileMapper.fromDto(profileDto);
+    }
 
     return ProfileModel.fromJson(response);
   }