+++ bricks/ui_kit_component/__brick__/example/lib/pages/{{name.snakeCase()}}_page.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:{{packageName}}/{{packageName}}.dart';
 
 class {{name.pascalCase()}}Page extends StatelessWidget {
   const {{name.pascalCase()}}Page({super.key});