+++ bricks/ui_kit_component/__brick__/lib/{{appName.snakeCase()}}_ui_kit.dart
@@ -1,3 +1,0 @@
-
-{{#hasEnum}}export 'src/components/{{name.snakeCase()}}/enums/{{prefix.snakeCase()}}_{{name.snakeCase()}}_enums.dart';{{/hasEnum}}
-export 'src/components/{{name.snakeCase()}}/{{prefix.snakeCase()}}_{{name.snakeCase()}}.dart';