+++ bricks/ui_kit_component/__brick__/lib/{{packageName}}.dart
@@ -1,0 +1,3 @@
+
+{{#hasEnum}}export '{{{componentPath}}}/enums/{{prefix.snakeCase()}}_{{name.snakeCase()}}_enums.dart';{{/hasEnum}}
+export '{{{componentPath}}}/{{prefix.snakeCase()}}_{{name.snakeCase()}}.dart';