+++ bricks/ui_kit_core/__brick__/lib/src/core/theme/{{prefix.lowerCase()}}_color_theme.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../../../{{name.snakeCase()}}_ui_kit.dart';
+import '../../../{{packageName}}.dart';
 
 class {{prefix.upperCase()}}ColorTheme extends ThemeExtension<{{prefix.upperCase()}}ColorTheme> {