+++ bricks/ui_kit_core/__brick__/lib/src/core/utils/component_init.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 
-import '../../../{{name.snakeCase()}}_ui_kit.dart';
+import '../../../{{packageName}}.dart';
 
 class {{prefix.upperCase()}}ComponentInit extends StatelessWidget {
   const {{prefix.upperCase()}}ComponentInit({super.key, required this.builder});