+++ bricks/ui_kit_core/CHANGELOG.md
@@ -15,3 +15,15 @@
   - Check if `rename` package is installed. If not, then it will run `dart pub global activate rename`
   - Run `rename` to update app name to "App Name UI Kit".
   - Remove all unused files.
+
+# 0.3.0+3
+
+- Improved prompts during skeleton generation:
+  - Added prompt for the number of colors to be generated.
+  - Added prompt for the number of text styles to be generated.
+  - Prompt users to provide information about each color and text style item.
+- Fixed an issue with the `change_app_package_name` process, which was not updating the app package name and bundle id correctly.
+- Renamed the example config files to `config.json` and removed the unnecessary `example` prefix.
+- Added a new `component-config.json` file that users can use to generate components using the UI Kit component brick with the command `mason make ui_kit_component -c component-config.json`. Users can customize the content of `component-config.json` according to their needs.
+- Implemented a post-generation process to fix code violations using `dart fix --apply`.
+- Added a final congratulatory message when the component is generated successfully.